--- fourtoon-cookie/src/components/layout/MainPageLayout/MainPageLayout.styled.ts
@@ -1,14 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#F7F7F7',
-    },
-    childrenContainer: {
-        position: 'relative',
-        flexGrow: 1,
-        height: "89%",
-        padding: 10,
-    }
-});
