+++ fourtoon-cookie/src/pages/DiaryTimelinePage/DiaryTimelinePage.tsx
@@ -3,8 +3,8 @@ import {FlatList} from "react-native";
 import Header from "./Header";
 import DiaryEmpty from "./DiaryEmpty";
 import DiaryComponent from "./DiaryComponent/DiaryComponent";
-import MainPageLayout from "../../components/layout/MainPageLayout/MainPageLayout";
-import { FOOTER_STATE } from "../../components/layout/MainPageLayout/Footer/Footer";
+import MainPageLayout from "../../components/layout/MainPageLayout";
+import { FOOTER_STATE } from "../../components/layout/MainPageLayout";
 import { useDiaries } from "../../hooks/server/diary";
 import { Diary } from "../../types/diary";
 import { AccountStatus } from "../../types/account";
